server/requirements.txt
fastapi[all]
uvicorn
pymongo[srv]
python-dotenv